Add data to c code

+++ c-svc/transform.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void addVectorToTimeSeriesObservation(TimeSeriesObservation *tso, double newVector[ACCELEROMETOR_VECTOR_DIM]);
+void addVectorToTimeSeriesObservation(TimeSeriesObservation *tso, int newVector[ACCELEROMETOR_VECTOR_DIM]);
 void transform(TimeSeriesObservation o, SampledObservation *out);
 
 #endif /* END OF TRANSFORM_H */
